Move createBricks to the game class
