+++ lib/index.js
@@ -21,10 +21,10 @@ getFromStorage();
 
 function generateBricks() {
   if (newGame.level === 1 || newGame.level === 2) {
-    let newBricks = bricks.createBricksLvlsOneTwo(40, newGame.level);
+    let newBricks = newGame.createBricksLvlsOneTwo(40, newGame.level);
     newBricks.forEach( brick => newGame.grabBricks(brick) );
   } else if (newGame.level === 3) {
-    let newBricks = bricks.createBricksLvlThree(54);
+    let newBricks = newGame.createBricksLvlThree(54);
     newBricks.forEach( brick => newGame.grabBricks(brick) );
   }
 }
